WIP: Initial commit
by
* Walking the AST of a file, and logging each interesting item
* Requesting references to that item from the LSP

AST walking may seem a little overly complex. We already have an LSP,
why not just ask it for the symbols it knows about using
textDocument/documentSymbols? We did start there, but missed the
selectionRange field and thought we couldn't find the name of the symbol
from the information provided. Any further justification would be a lie,
but I also have nothing against using treesitter, so we leave it the way
it is.

AST walking has been implemented in a language agnostic way. We have
provided a set of rules that are listed in the input config that specify
which hierarchical elements we are interested in, and how to resolve
their names. An example config has been provided in res/config.json

The amount of code is increasing, too much for everything to continue to
sit in main. An attempt to use a set of common types has been made,
nothing lsp should leak out of the lsp section, nothing treesitter
should leak out of the treesitter section.

Initial commit

Write a basic LSP client that can be used to query references in a given
file/line/col